systemd: respect USE_LDCONFIG when setting default PACKAGECONFIG
Avoid trying to call ldconfig at run-time in distros which force
USE_LDCONFIG to 0 (and therefore don't provide ldconfig support in
the target rootfs).

Note that this approach of checking USE_LDCONFIG only applies to
OE 2.2 and earlier. In OE 2.3, the USE_LDCONFIG global variable was
replaced by the "ldconfig" distro feature.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>