glibc: disable ld.so.cache lookup based on USE_LDCONFIG
USE_LDCONFIG defaults to 1 but if forced to 0 (e.g. from a disto
config), then patch glibc to disable the hidden "use_ldconfig" config
option, which is otherwise always enabled for Linux targets.

Disabling use_ldconfig forces the dynamic linker only pick up
libraries from the library path and rpath and ignore /etc/ld.so.cache

This patch is the equivalent of always passing the --inhibit-ldcache
option to ld.so:

  https://sourceware.org/git/?p=glibc.git;a=commit;h=73d65cc3784b1413b017fed31550ec11e717e45b

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>